+++ chromance-firmware/src/animations/chromance/ripplePool.h
@@ -16,15 +16,15 @@ namespace Chromance
              * @param animationId The ID of the animation that is requesting to claim a ripple from the pool
              * @return A ripple that was successfully claimed by the caller
             */
-            Ripple* Claim(uint8_t animationId);
+            Ripple* Claim(int32_t animationId);
             /**
              * Get a ripple by index
              * @param index The index of the ripple 
              * @return A ripple
             */
-            Ripple* Get(uint8_t index);
+            Ripple* Get(uint32_t index);
 
-            static const uint8_t NumberOfRipples = 30U;
+            static const uint32_t NumberOfRipples = 30U;
 
         private:
