+++ chromance-firmware/src/animations/randomPulseAnimation.h
@@ -9,12 +9,12 @@ namespace Chromance
     {
         public:
 
-            RandomPulseAnimation(uint8_t id, RipplePool* ripplePool, Logger* logger);
+            RandomPulseAnimation(int32_t id, RipplePool* ripplePool, Logger* logger);
 
             void Start();
 
         private:
-            uint8_t lastPulseNode;
+            int32_t lastPulseNode;
     };
 }
